% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_screening.R
\name{Cluster_screening}
\alias{Cluster_screening}
\title{Post-processing, called screening, of the detected change-points}
\usage{
Cluster_screening(Tmu, p_val = 0.05, MaxDist = 80, detail = NULL)
}
\arguments{
\item{Tmu}{the segmentation results obtained from the Segmentation function}

\item{p_val}{a number from 0 to 1 indicating the threshold for p-value}

\item{MaxDist}{the maximal number of days between change-points used to determinate the cluster. Default is 80}

\item{detail}{any object indicating to have detail of test as output. Defaul is NULL, do not include the details.}
}
\value{
A list of four variables (RemoveData, UpdatedCP, ChangeCP, and detail):
\itemize{
\item \code{UpdatedCP}: The change-points remaining after screening.
\item \code{RemoveData}: A data frame containing the beginning and end positions (time index) of the segments to be deleted after filtering.
\item \code{ChangeCP}: "Yes" or "No", indicating whether the list of change-points changed after cluster screening.
\item \code{detail}: A data frame containing the results of the significance test of the mean difference before and after the cluster, including the following columns:
\itemize{
\item \code{mu_L}, \code{mu_R}: Means of the segments before and after the cluster.
\item \code{se_L}, \code{se_R}: Standard errors of the means of the segments before and after the cluster.
\item \code{np_L}, \code{np_R}: Number of points before and after the cluster.
\item \code{tstat}, \code{pval}, \code{signif}: t-values, p-values, and significance codes of the test.
}
}
}
\description{
(1) detection of cluster of change-points (formed by consecutive change-points separated by a maximum number of days)
(2) testing the variation of mean before and after the cluster
}
\details{
The post-processing rule is the following: the cluster is removed if the mean difference is unsignificant whereas the cluster is replaced by a change-point in the middle if the mean difference is significant. In both cases, some data will be to removed in the next step of the analysis
}
