% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSeg.R
\name{PlotSeg}
\alias{PlotSeg}
\title{Plot the time series with segmentation results: change-point positions, estimatedmean , with the monthly variance and function (if specified in the model)}
\usage{
PlotSeg(
  OneSeries,
  SegRes,
  FunctPart = TRUE,
  RemoveData = NULL,
  Metadata = NULL,
  Validated_CP_Meta = NULL,
  labelx = "Date",
  labely = "signal"
)
}
\arguments{
\item{OneSeries}{a data frame, with size n x 2, containing the signal with n points and the dates in Date format. The names of the 2 columns are thus signal and date}

\item{SegRes}{A list containing the results of the segmentation process.
This may include:
\itemize{
\item \code{Tmu}: the segmentation results
\item \code{FitF}: the estimated functional part. If the function is not taking into account in the model, \code{FitF} is FALSE
\item \code{MonthStd}: the estimated standard deviations of each month
}}

\item{FunctPart}{a boolean indicating if the functional part is taking into account in the model. Default is TRUE and note that if \code{FunctPart=FALSE}, only a segmentation is performed}

\item{RemoveData}{a data frame including the beginning and the end positions (time index) of the segments to be deleted after filtering}

\item{Metadata}{a data frame with two columns: $date that is the known change in same format as date of \code{OneSeries} and the $type that is the type the change}

\item{Validated_CP_Meta}{the results from the Validation function. Default is NULL}

\item{labelx}{a name representing the label of the x-axis}

\item{labely}{a name representing the label of the y-axis}
}
\value{
a object of class \code{\link[ggplot2]{ggplot}} representing the
time series plot with segmentation and validation results overlaid.
}
\description{
Plot the time series with segmentation results: change-point positions, estimatedmean , with the monthly variance and function (if specified in the model)
}
