% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{Segmentation}
\alias{Segmentation}
\title{Segmentation of time series}
\usage{
Segmentation(
  OneSeries,
  lmin = 1,
  Kmax = 30,
  selectionK = "BM_BJ",
  FunctPart = TRUE,
  selectionF = FALSE
)
}
\arguments{
\item{OneSeries}{a data frame, with size n x 2, containing the signal with n points and the dates in Date format. The names of the 2 columns are thus signal and date}

\item{lmin}{the minimum length of the segments. Default is 1}

\item{Kmax}{the maximal number of segments (must be lower than n). Default is 30}

\item{selectionK}{a name indicating the model selection criterion to select the number of segments K (\code{mBIC}, \code{Lav}, \code{BM_BJ} or \code{BM_slope}). \code{"none"} indicates that no selection is claimed and the procedure considers \code{Kmax} segments or \code{Kmax}-1 changes. Default is \code{"BM_BJ"}}

\item{FunctPart}{a boolean indicating if the functional part is taking into account in the model. Default is TRUE and note that if \code{FunctPart=FALSE}, only a segmentation is performed}

\item{selectionF}{a boolean indicating if a selection on the functions of the Fourier decomposition of order 4 is performed. The level of the test is by default 0.001. Default is FALSE}
}
\value{
A file containing
\itemize{
\item \code{Tmu} that is a data frame containing the estimation of the segmentation parameters: for each segment, we get the beginning and the end positions (time index), the estimated mean (mean), the standard deviation of the mean (se) and the number of points (np)
\item \code{FitF} that corresponds to the estimation of the functional part. If \code{FunctPart=FALSE}, \code{FitF} is FALSE
\item \code{CoeffF} that corresponds to the estimation of the coefficients of the Fourier decomposition. The vector contains 8 coefficients if \code{selectionF=FALSE} or as many coefficients as the number of selected functions if \code{selectionF=TRUE}. If \code{FunctPart=FALSE}, \code{CoeffF} is FALSE
\item \code{MonthVar} that corresponds to the estimated variances of each fixed interval (each month)
\item \code{SSR} that corresponds to the Residuals Sum of Squares for \code{Kmax} segments
}
}
\description{
fit a segmentation in the mean model by taken into account a functional part and a monthly variance
}
\details{
The function performs the segmentation of one GNSS series. The considered model is such that: (1) the average is composed of a piece wise function (changes in the mean) with a functional part and (2) the variance is heterogeneous on fixed intervals. By default the latter intervals are the months.
The inference procedure consists in two steps. First, the number of segments is fixed to \code{Kmax} and the parameters are estimated using the maximum likelihood procedure using the following procedure: first the variances are robustly estimated and then the segmentation and the functional parts are iteratively estimated. Then the number of segments is chosen using model selection criteria. The possible criteria are \code{mBIC} the modified BIC criterion, \code{Lav} the criterion proposed by Lavielle, \code{BM_BJ} and \code{BM_slope} the criteriain which the penalty constant is calibrated using the Biggest Jump and the slope.
\itemize{
\item The data is a data frame with 2 columns: $signal is the signal to be segmented and $date is the date. The date will be in Date format.
\item The function part is estimated using a Fourier decomposition of order 4 with \code{selectionF=FALSE}. \code{selectionF=TRUE} consists in selecting the significant functions of the Fourier decomposition of order 4
\item If \code{selectionK="none"}, the procedure is performed with \code{Kmax} segments.
\item Missing data in the signal are accepted.
}
}
