% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_CP.R
\name{Test_CP}
\alias{Test_CP}
\title{Test of each change-point (CP)}
\usage{
Test_CP(Tmu, alpha = 0.05, detail = FALSE)
}
\arguments{
\item{Tmu}{the segmentation results obtained from the Segmentation function}

\item{detail:}{if TRUE the output contains a $detail field with additional information on the test. Default is FALSE.}

\item{alpha:}{significance level of the test (value between 0 and 1). Default is 0.05.}
}
\value{
\itemize{
\item \code{UpdatedCP}: The change-points remaining after testing.
\item \code{RemoveData}: A data frame containing NA to be consistent with the results of the screening.
\item \code{ChangeCP}: "Yes" or "No", indicating whether the list of change-points changed after the test.
\item \code{detail}: A data frame containing the results of the test of the mean difference before and after the cluster, including the following columns:
\itemize{
\item \code{mu_L}, \code{mu_R}: Means of the segments before and after the change-point.
\item \code{se_L}, \code{se_R}: Standard errors of the means of the segments before and after the change-point.
\item \code{np_L}, \code{np_R}: Number of points before and after the change-point.
\item \code{tstat}, \code{pval}, \code{signif}: t-values, p-values, and significance codes of the test.
}
}
}
\description{
Method:
(1) test of the change in mean before and after the cluster with significance level alpha
(2) removal of all the CPs in the cluster if the change in mean is unsignificant
}
