% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateTimeSeries.R
\name{UpdateTimeSeries}
\alias{UpdateTimeSeries}
\title{Post-processing of the time series after cluster screening}
\usage{
UpdateTimeSeries(OneSeries, RemoveData)
}
\arguments{
\item{OneSeries}{A time series data frame with 2 columns, $signal and $date, each of size n x 1}

\item{RemoveData}{A data frame containing the beginning and end positions (time index) of the segments to be deleted in the data.}
}
\value{
UpdatedSeries The time series after scrrening
}
\description{
Method:
(3) replaces all the data points in the clusters detected by Cluster_screening() by NA
}
