% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{Validation}
\alias{Validation}
\title{Validation of change-points using metadata}
\usage{
Validation(OneSeries, Tmu, MinDist = 62, Metadata)
}
\arguments{
\item{OneSeries}{a data frame, with size n x 2, containing the signal with n points and the dates in Date format. The names of the 2 columns are thus signal and date}

\item{Tmu}{the segmentation results}

\item{MinDist}{the minimum distance from change-points to known changes to be validated}

\item{Metadata}{a data frame with two columns: $date that is the known change in same format as date of \code{OneSeries} and the $type that is the type the change}
}
\value{
a data frame with four columns :
\itemize{
\item \code{CP} that is the change-points expressed in date to be validated
\item \code{closestMetadata} that is the closest date of a known change
\item \code{Distance} that is the distance between \code{CP} and \code{closestMetadata}
\item \code{type} that is the type of the known change
\item \code{valid} that takes the value 1 if the change-point is validated, 0 otherwise using a distance of \code{MinDist}
}
}
\description{
Validation of change-points using metadata
}
