% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_screening.R
\name{Cluster_screening}
\alias{Cluster_screening}
\title{Post-processing, called screening, of the detected change-points}
\usage{
Cluster_screening(Tmu, MaxDist = 80)
}
\arguments{
\item{Tmu}{the segmentation results obtained from the Segmentation function}

\item{MaxDist}{the maximal number of days between change-points used to determinate the cluster. Default is 80}
}
\value{
list of two variables (RemoveData and UpdatedCP) :
\itemize{
\item \code{UpdatedCP} that is the change-points remaining after screening
\item \code{RemoveData} that is a data frame including the beginning and the end positions (time index) of the segments to be deleted after filtering
}
}
\description{
(1) detection of cluster of change-points (formed by consecutive change-points separated by a maximum number of days)
(2) testing the variation of mean before and after the cluster
}
\details{
The post-processing rule is the following: the cluster is removed if the mean difference is unsignificant whereas the cluster is replaced by a change-point in the middle if the mean difference is significant. In both cases, some data will be to removed in the next step of the analysis
}
