% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_screening.R
\name{Cluster_screening}
\alias{Cluster_screening}
\title{Post-processing of the segmentation results aiming to remove clusters of change-points (CPs)}
\usage{
Cluster_screening(Tmu, alpha = 0.05, MaxDist = 80, detail = FALSE)
}
\arguments{
\item{Tmu}{the segmentation results obtained from the Segmentation function}

\item{MaxDist}{the maximal number of days between change-points within a cluster. Default is 80.}

\item{detail:}{if TRUE the output contains a $detail field with additional information on the test. Default is FALSE.}

\item{alpha:}{significance level of the test (value between 0 and 1). Default is 0.05.}
}
\value{
\itemize{
\item \code{UpdatedCP}: The change-points remaining after screening.
\item \code{RemoveData}: A data frame containing the beginning and end positions (time index) of the segments to be deleted after filtering.
\item \code{ChangeCP}: "Yes" or "No", indicating whether the list of change-points changed after cluster screening.
\item \code{detail}: A data frame containing the results of the significance test of the mean difference before and after the cluster, including the following columns:
\itemize{
\item \code{mu_L}, \code{mu_R}: Means of the segments before and after the cluster.
\item \code{se_L}, \code{se_R}: Standard errors of the means of the segments before and after the cluster.
\item \code{np_L}, \code{np_R}: Number of points before and after the cluster.
\item \code{tstat}, \code{pval}, \code{signif}: t-values, p-values, and significance codes of the test.
}
}
}
\description{
Method:
(1) detection of cluster of CPs, i.e. consecutive CPs closer than MaxDist days
(2) test of the change in mean before and after the cluster with significance level alpha
(3) removal of all the CPs in the cluster if the change in mean is unsignificant
(4) add a new change-point in the middle of the cluster if the change in mean is significant.
}
\details{
Note: it is recommended to replace the data inside the cluster(s) by NA in the general homogenization process
}
