% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSeg.R
\name{PlotSeg}
\alias{PlotSeg}
\title{Plot the time series with segmentation results: change-point positions, estimatedmean , with the monthly variance and function (if specified in the model)}
\usage{
PlotSeg(
  OneSeries,
  SegRes,
  FunctPart = TRUE,
  RemoveData = NULL,
  Metadata = NULL,
  Validated_CP_Meta = NULL,
  labelx = "date",
  labely = "signal",
  title = "Segmented time series"
)
}
\arguments{
\item{OneSeries}{a data frame, with size n x 2, containing the signal with n points and the dates in Date format. The names of the 2 columns are thus signal and date}

\item{SegRes}{result from the segmentation funtion.}

\item{FunctPart}{a boolean, specifies if the functional part should be plotted. Default is FALSE.}

\item{RemoveData}{a data frame including the beginning and the end positions (time index) of the segments to be removed.}

\item{Metadata}{a data frame with two columns: $date (in date format) and $type (label in string format)}

\item{Validated_CP_Meta}{the results from the Validation function. Default is NULL.}

\item{labelx}{a string for the label of the x-axis}

\item{labely}{a string for the label of the y-axis}
}
\value{
a object of class \code{\link[ggplot2]{ggplot}} representing the
time series plot with segmentation and validation results overlaid.
}
\description{
Plot the time series with segmentation results: change-point positions, estimatedmean , with the monthly variance and function (if specified in the model)
}
