% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{Segmentation}
\alias{Segmentation}
\title{Segmentation of time series by Penalized Maximum Likelihood}
\usage{
Segmentation(
  OneSeries,
  lmin = 1,
  Kmax = 30,
  selectionK = "BM_BJ",
  FunctPart = TRUE,
  selectionF = FALSE,
  initf = "ascendent"
)
}
\arguments{
\item{OneSeries}{is a time series data frame with 2 columns, $signal and $date, each of size n x 1, n is the number of days of the time series.
Note: the $date variable should be continous. If the original data has gaps, NAs should be added at the corresponding dates.}

\item{lmin}{is the minimum length of the segments. Default value is 1.}

\item{Kmax}{is the maximal number of segments (Kmax < n). Default value is 30.
Note: with \code{BM_slope}, \code{Kmax} must be larger than or equal to 10.}

\item{selectionK}{specifies the penalty criterion used for the model selection (selection of the number of segments K <= Kmax).
Options are: \code{"none"}, \code{"mBIC"}, \code{"Lav"}, \code{"BM_BJ"} or \code{"BM_slope"}). Default is \code{"BM_BJ"}.
If \code{selectionK = "none"}, the model is estimated with \code{K = Kmax}.
If \code{selectionK = "All"}, the results for the four possible criteria are given.}

\item{FunctPart}{specifies if the functional part (Fourier series of order 4) should be included in the model (\code{FunctPart=TRUE}) or not (\code{FunctPart=FALSE}). Default is TRUE.
Note: with \code{FunctPart=TRUE} the algorithm estimates the functional and the segmentation parameters in an iterative way.
with \code{FunctPart=FALSE}, only one segmentation is performed.
If the functional part is unnecessary, \code{FunctPart=FALSE} can be much faster.}

\item{selectionF}{is used to select only significant coefficients of the Fourier series when \code{FunctPart=TRUE}. Default is FALSE.}

\item{initf}{refers how the functional part is initialized in the iterative inference procedure.
Options: \code{"each"} means that the functional part is estimated at each K,
\code{"ascendent"} means that the initial functional part for K is the solution obtained for K-1,
\code{"descendent"} means that the initial functional part for K is the solution obtained for K+1,
\code{"both"} means that the \code{"ascendent"} is used first for K from 1 to Kmax, then \code{"descendent"} is tested from Kmax/2 to 1 (if the solution is better, it replaced the \code{"ascendent"} solution.
Default is \code{"ascendent"}.}
}
\value{
\itemize{
\item \code{Tmu} is a data frame containing the segmentation results, with 5 columns and a number of lines equal to the number of segments of the time series.
The columns are: \code{$begin, $end, $mean, $se, $np}. They represent the date index (integer) of begin and end of each segment, the estimated mean of the segment (\code{mean}) and its standard error (\code{se}), and the number of "valid" points (\code{np}), i.e. non-NA $signal values in the segment.
\item \code{FitF} is the functional part predicted from the estimated Fourier coefficients, a numeric vector of size n x 1. Note: if \code{FunctPart=FALSE}, \code{FitF} is FALSE.
\item \code{CoeffF} is the vector of coefficients of the Fourier series, a numeric vector of size 1 x 8 if \code{selectionF=FALSE}.
Note: If \code{selectionF=TRUE} the size of \code{CoeffF} correspods to the number of selected coefficients.
If \code{FunctPart=FALSE}, \code{CoeffF} is FALSE.
\item \code{MonthVar} contains the estimated monthly variances, a numeric vector of size 1 x 12.
\item \code{SSR} is the Sum of Squared Residuals of the fit.
\item \code{SSR_All} is the Sum of Squared Residuals for the fit for all K=1,...,Kmax
}
If \code{selectionK="All"}, the outputs are each a list containing the corresponding results obtained for all the model selection criteria
}
\description{
Method: estimate the position of change-points (CPs) in the mean superposed with a global periodic function (Fourier series of order 4) and IID noise with variance changing over fixed intervals.
}
\details{
The current version assumes that: i) the time series is given with a daily time resolution, ii) the functional has a fundamental period of 1 year, and iii) the variances are chaning on monthly intervals.

The theoretical basis of the segmentation method developped by Quarello (2020) and published in Quarello et al., (2022).
The inference procedure consists in three steps:
\enumerate{
\item The monthly variances are estimated using a robust method following Bock et al. (2020).
\item The segmentation parameters (change-point positions and segments' means) and the functional parameters (Fourier coefficients) are estimated iteratively, for a fixed number of segments \code{K = 1..Kmax}.
The estimation method is based on maximum likelihood with known variance (from step 1).
The segmentation and functional parameters are estimated separately, which allows to use the Dynamical Programming algorithm for the search of the optimal change-point positions and segment means.
\item The optimal number of segments is obtained by model selection using one of the following penalty criteria:
\itemize{
\item \code{mBIC}, the modified Bayesian Inference Criterion proposed by Zhang and Siegmund (2007),
\item \code{Lav}, the criterion proposed Lavielle (2005),
\item \code{BM_BJ} and \code{BM_slope}, the criteria proposed by Birgé and Massart (2001), where the penalty constant is calibrated using the Biggest Jump and the slope, respectively.
}
}

Note: by convention, the position of a change-point refers to the last point in a segment (\code{Tmu$end}).
}
\references{
Birgé, L.; Massart, P. (2001) Gaussian model selection. J. Eur. Math. Soc. 2001, 3, 203–268, DOI 10.1007/S100970100031.

Bock, O.; Collilieux, X.; Guillamon, F.; Lebarbier, E.; Pascal, C. (2020) A breakpoint detection in the mean model with heterogeneous variance on fixed time intervals. Statistics and Computing 2020, 30, 195–207. https://doi.org/10.1007/s11222-019-09853-5.

Lavielle, M. (2005) Using penalized contrasts for the change-point problem. Signal Processing 2005, 85, 1501–1510.

Quarello, A. (2020) A. Development of New Homogenisation Methods for GNSS Atmospheric Data. Application to the Analysis of Climate Trends and Variability. Ph.D. Thesis, Sorbonne Universite, Paris, France, 2020

Quarello, A., Bock, O. & Lebarbier, E. (2022) GNSSseg, a statistical method for the segmentation of daily GNSS IWV time series. Remote Sensing, 14(14), 3379. Available from: https://doi.org/10.3390/rs14143379

Zhang, N.R.; Siegmund, D.O. (2007) A Modified Bayes Information Criterion with Applications to the Analysis of Comparative Genomic Hybridization Data. Biometrics 2007, 63, 22–32.
}
