% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdatedParametersForFixedCP.R
\name{UpdatedParametersForFixedCP}
\alias{UpdatedParametersForFixedCP}
\title{update the segmentation parameters from the screening result}
\usage{
UpdatedParametersForFixedCP(
  OneSeries,
  ResScreening,
  FunctPart = TRUE,
  selectionF = FALSE
)
}
\arguments{
\item{OneSeries}{a data frame, with size n x 2, containing the signal with n points and the dates in Date format. The names of the 2 columns are thus signal and date}

\item{ResScreening}{the output of the Cluster_screening function}

\item{FunctPart}{a boolean indicating if the functional part is taking into account in the model. Default is TRUE and note that if \code{FunctPart=FALSE}, only a segmentation is performed}

\item{selectionF}{a boolean indicating if a selection on the functions of the Fourier decomposition of order 4 is performed. The level of the test is by default 0.001. Default is FALSE}
}
\value{
A file containing
\itemize{
\item \code{Tmu} that is a data frame containing the estimation of the segmentation parameters: for each segment, we get the beginning and the end positions (time index), the estimated mean (mean), the standard deviation of the mean (se) and the number of points (np)
\item \code{FitF} that corresponds to the estimation of the functional part. If \code{FunctPart=FALSE}, \code{FitF} is FALSE
\item \code{CoeffF} that corresponds to the estimation of the coefficients of the Fourier decomposition. The vector contains 8 coefficients if \code{selectionF=FALSE} or as many coefficients as the number of selected functions if \code{selectionF=TRUE}. If \code{FunctPart=FALSE}, \code{CoeffF} is FALSE
\item \code{MonthVar} that corresponds to the estimated variances of each fixed interval (each month)
}
}
\description{
update the segmentation parameters from the screening result
}
