% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{Validation}
\alias{Validation}
\title{Validation of change-points using metadata}
\usage{
Validation(OneSeries, Tmu, MaxDist = 62, Metadata)
}
\arguments{
\item{OneSeries}{is a time series data frame with 2 columns, $signal and $date, each of size n x 1}

\item{Tmu}{is a data frame containing segmentation results (see segmentation.R)}

\item{MaxDist}{is the maximum distance (number of valid days) between a change-point and the nearest metadata to be valid}

\item{Metadata}{is a data frame with two columns, $date and $type, they represent changes at the station which observed the OneSeries time series}
}
\value{
\itemize{
\item \code{CP} are the dates of the change-points contained in \code{Tmu}
\item \code{closestMetadata} are the nearest dates found in \code{Metadata}
\item \code{Distance} is the distance between \code{CP} and \code{Metadata}
\item \code{type} is the corresponding type of change from \code{Metadata}
\item \code{valid} is the validation result: 1 if \code{Distance} is smaller than \code{MaxDist}, 0 otherwise
}
}
\description{
Validation of change-points using metadata
}
